refactor: renaming and fixing UI bugs
